Will/account deletion (#35)

+++ screens/modals/ChangeInfoScreen.tsx
@@ -48,6 +48,7 @@ export default function ChangeInfoScreen({ route, navigation }: Props) {
     const [message, setMessage] = useState<string | null>(null);
 
     const allowChange = state.matches("Information Valid") && state.context.infoChanged;
+    console.log(state.matches("Information Valid"), state.context.infoChanged);
 
     useEffect(() => {
         if (!userID || !userInfo) return;
@@ -63,7 +64,7 @@ export default function ChangeInfoScreen({ route, navigation }: Props) {
         });
     }, [name, major, grad, userID, send, pronouns]);
 
-    const onConfirm = async () => {
+    const onConfirm = () => {
         Alert.alert("Confirm Action", "Are you sure you want to make these changes?", [
             {
                 text: "Cancel",
@@ -79,13 +80,24 @@ export default function ChangeInfoScreen({ route, navigation }: Props) {
         ]);
     };
 
-    const onDelete = async () => {
-        if (!userID) {
-            setMessage("Error: no user found.");
-            return;
-        }
-        const res = await deleteAccount(userID);
-        if (res.type === MessageType.error) setMessage(res.message);
+    const onDelete = () => {
+        Alert.alert("Confirm Action", "Are you sure you want to delete your account?", [
+            {
+                text: "Cancel",
+                onPress: () => {},
+            },
+            {
+                text: "Confirm",
+                onPress: async () => {
+                    if (!userID) {
+                        setMessage("Error: no user found.");
+                        return;
+                    }
+                    const res = await deleteAccount(userID);
+                    if (res.type === MessageType.error) setMessage(res.message);
+                },
+            },
+        ]);
     };
 
     return (
